Two very minor changes.
TODO - I'm not sure if these changes are actually necessary or some litecoin/gamerscoin hangover, we may be able to dump them entirely. (Check this)
It seems suspicious that they are both the same value (50000) replaced with 9924 - could they be from a bad s&r or something? Track the orgins of this down and have a closer look...