Various small bug fixes.
Merge a bug fix from user 'mezrin' for 'dumpwallet' RPC command.
Fix a bug in RPC command 'gettransaction' where transaction would not show if 'rpconlylistsecuredtransactions' flag was set to false.
Obscure the genesis block message as a hex string instead of plain text (to hide the fact that it has a grammar error that some people like to point out)