Fold build system changes back into codebase.
Previously the build system automatically changed various items in the codebase in order to keep the main codebase as similar to bitcoin as possible (as a strategy to try keep merging simple) - at this point after 2.0 release we have however diverged from Bitcoin so far that this strategy is no longer valid.
So we now just implement this changes into the main codebase, and simplify the build system.