Fix issues arouind phase 4 upgrade behaviour and other minor PoW2 related changes.
Force nVersion to equal nVersionPoW2Witness of embedded witness during phase 3.
Simplify phase 4 upgrade code to rely on the above (it can use nVersion now instead of needed to extract the embedded witness)
Rename IsWitnessEnabled to IsSegSigEnabled for clarity.
Turn validation of witness signature back on (it was temporarily disabled when GetWitness function was upgraded to use the CGetWitnessInfo struct)
Simplify IsSegSigEnabled to simply check if the previous block (must be the previous block - as in phase4 the previous block is always witnessed) has nVersionPoW2Witness set (witnessed) or not instead of relying on much more expensive and finnicky phase4 checks - this fixes several bugs in the process (related to version bit cachine and phase4 check being called "too early" in the sync process with blocks not yet connected to chain) as well as improves performance.