New way for for header sync progress
Number of expected headers was previously based on the time passed since
the time of the header tip and the target block spacing. This didn'work
because:
- with reverse headers the header tip stays at height 0 for all the
headers up to the last checkpoint.
- At height 0 we have the Genesis which has a weird date (Jan 1, 2002).
- The actual block spacing over all of the history is off by alomst 20%
The new solution uses a probably block height based on what peers
report.
The header sync progress message is now only given when getting headers
over the network, ie. not when the tip advances because blocks are read
from disk etc.