User agent http API requests same as in version message in p2p, set amountType in Nocks request
Nocks had a workaround in their API based on our user agent. By improving and versioning our user agent we automatically bypass the workaround.
Also we trigger the correct API by passing the amountType which as now been fixed on their end such that the quote and transaction amounts match.