Extend recovery phrase with (optional) birth number.
When recovering the date to start recovering is computed from the birth number, greatly enhancing recovery speed.
This is important for SPV mode as otherwise a recovery will always need to start at the genesis, effectively destroying most of the SPV improvement when a recovery is needed.
The birth number is computed by taking the age of the wallet since the genesis block expressed in weeks.
If no birth number is used the recovery acts as it used to starting from the genesis. (+1 squashed commit)
Squashed commits:
[c5b7a755d] Extend recovery phrase with (optional) birth number.

When recovering the date to start recovering is computed from the birth number, greatly enhancing recovery speed.
This is important for SPV mode as otherwise a recovery will always need to start at the genesis, effectively destroying most of the SPV improvement when a recovery is needed.
The birth number is computed by taking the age of the wallet since the genesis block expressed in weeks.
If no birth number is used the recovery acts as it used to starting from the genesis.