New RPC command "emptymempool" - this is useful for various debugging situations.
And may also be a useful tool to have on hand should a mining pool find themselves stuck with an invalid mempool transaction somehow that causes getblocktemplate to create invalid blocks.
Obviously the code ideally prevents this from happening; however its good to have the right tools to fall back on if something goes wrong.