Move Djinni generated code for iOS in tree and add into repo
This makes the build a bit simpler also for iOS simulator.
Putting generated code into the repo might seem illogical, though it really makes the build easier for using the Unity library.
Also already part of it for Android is already included and the build system already references the files that will be generated.