Fix crash in pure partial sync when started from height 0
When pure partial sync was started from height 0 a later sessions could crash (would manifest on a non-connecting incoming block announcement).
Problem was that when starting from height 0 the main tree (and its validation bits) would grow on the incoming blocks requested for spv, but on later sessions the block index would be cleared possibly wiping blocks that were part of the main tree and that pindexBestHeader and other were already refering to.
This is now solved by first clearing the block index (and wiping status bits in pure partial sync) BEFORE initializing pindexBestHeader etc and by not growing the main tree (and its valid bits) when adding block indexes in pure partial sync.