CORE: Introduce "getBlockFilterBirthAndRanges" to read back and process the information from "dumpfiltercheckpoints"
Function takes a range of addresses to compare against as well as the current suspected wallet birth date
For the range of addresses in determines which ranges in the chain are of potential interest and returns these ranges, the sync code can then use these ranges to speed up sync.
Ranges -before- the suspected birth date are discarded as false positives
Where possible the birth date is advanced to the first range -after- the suspected birth date, in some instances (luck based due to false positives) this can allow the birth date to move quite far forward and allow less headers to fetch.

Reads three 'env' args 'spvstaticfilterfile', 'spvstaticfilterfileoffset' and 'spvstaticfilterfilelength' to control where the data comes from.
For platforms with direct file access only 'spvstaticfilterfile' is needed, for platforms like android the others can be used to read into a portion of a larger file (APK)