ANDROID: BUILD: attempt to fix randomly missing libgulden_unity_jni.so when building
Has been happening for a long time already (months) every now and then, no cause found so far.
When it happened cleaning the build folder solved it (which is really annoying).
Now the same issue also appeared when building the Android Unity lib on macOS. Have been hunting it for quite a while.
Not 100% sure that forcing this -shared flag fixes it, but so far I did not had any occurences (but they would be many builds apart sometimes).