CORE: Fix a rare witnessing issue when no proper rescan has occured.
"CWallet::SignTransaction" made some assumptions that the inputs for the transaction we are signing are always wallet transactions.
In the event that a witness account is imported into a wallet and the rescan interrupted then we don't have the funding (or other witness related transactions) in the wallet.
As a result we can/were winding up in a situation where we tried to sign blocks as the witness, but the signing failed due to the assumption about wallet transactions failing.
This despite the fact that we are actually perfectly capable of signing such a block...

This commit fixes this by changing the behaviour of "CWallet::SignTransaction" in such a way that it will work even for non wallet inputs (so long as it can locate the input from the utxo or block it is meant to be in)