Fix witness statistics value labels colors for macOS dark theme
When using dark theme on macOS (Mojave+) the standard text color is white, therefore these labels become invisible on a white background.
On the official releases this is so far not visible because because they are compiled using an older (pre Mojave) SDK and run in un-themed compatibility mode.
Note that when building native on macOS using a newer SDK also use a newer Qt that is Mojave aware (Qt 5.12+), else more UI will be distorted.