accepts address for destination in rpc setwitnessrewardscript
The help had an inconsistency in param naming, using address_or_script and pubkey_or_script to refer to the same thing, suggesting that it was originally the intention to accept addresses too.
Also it is far easier to come up with an address for the destination than a hex encoded script or pubkey.