QA: CORE: Potential BOOST_SCOPE_EXIT race condition fix
Fix a potential race condition around various BOOST_SCOPE_EXIT uses inside SIGMA, under certain optimisation conditions. Add extra scope guards to ensure the allocation is not reordered.
It is unclear whether this was an actual problem or not, but better to just be sure.
This might be the cause of the intermittent "double free" crashes some miners have been seeing, but this remains to be seen.