UNITY: ELECTRON: Change the order of paramater setting
App specific paramaters (e.g. -datadir) passed from specific frontends (e.g. electron) were being set before actually reading the .conf file, this resulted in a situation where various .conf settings may be ignored in favour of the hardcoded ones.
In the case of "Gulden lite" this meant that "datadir" in the Gulden.conf was ignored.
We shuffle the order of doing things around to prevent this.