CORE: Remove "undo.nHeight == 0" ApplyTxInUndo code
This code exists to upgrade from an older format, but novo has never had this format.
And its clashing wi th the fact that we have "genesis transactions" and causing false "db corruption" detection issues.

"The cause of the problem is now known, its rather technical, its to do with some of the premine transactions in the burn still being from the genesis block (block 0).
When the software does certain things it sometimes "disconnects" a block, and when it does that it needs to look up the transactions in something called an "undo database"; at some point the bitcoin people decided to assign a special meaning to "0" in this database for an upgrade they did to it, because they had no transactions in block 0 this wasn't a problem for them - its unfortunately triggered a problem for us.
The good news is the fix is probably simple, I just need to test it a bit."