BUG: Patch out DUST_HARD_LIMIT check
On Gulden we technically allow relaying of all dust (even if it doesn't make sense fee wise) so it doesn't make sense (and leads to some confusion with `checkwalletagainstutxo` if we don't consider these part of the wallet.
If we want to not count these as part of the balance or similar we should deal with this in some other way.